refactor: toplevel package to /internal package
