+++ internal/debug_print.go
@@ -1,4 +1,4 @@
-package selfish
+package internal
 
 import (
 	"encoding/json"
